Fix issue 14 18 (#38)
* - Fixes 14 and 18 by adding in inputs for solar radiation, date and time fields.
- Fixes 32 by adding Multistation support for plots and tables supported for fwi87 and fwi25
- TODO: fix timezones for sunrise,sunset and other timestamp fields

* Finally sorted out all the timezone bs.
ToDO:
Check plot data selection.
Sort out daily fwi 12:00 stuff (issue 19)

* AH! Fixed the times for real now. What a nightmare. Why can't things just be timezone aware or all in UTC.

* Okay, now for real, times are sorted.
The plots work as I expect, the tables look nice and so does the modal, and the helpers and tool tips.
But now it's super slow so will try to look into that.
Fixes #9
Fixes #14
Fixes #18
Fixes #12
Fixes #24
I think it fixes number 11,  27 but not sure.

* Hmm, can't decide if this is better or not but pushing it up.
Might be better to revert the plot code. But if not we need to fix the faceting and labelling.
Otherwise the next thing to work on is how slow the app is and that there's no indication that the button was pressed and it's now thinking.

* Everything but the timing of the spinners now.

* Timing fixed!

* Getting closer

* Finally found where the warning was coming from in the tables.