Dev (#41)
* Fixes #16 (#25)

Closes #16

* Allows user to provide or overwrite a latitude and longitude value. (#26)

* Issue 15 13 (#30)

* Moving to modular structure.
Import of csv working, update of lat/long and datetime from file now working.

* Got the timezone selection/detection working now.

* Minor update to mapping columns

* Translations now work.

* Mapping updated to include column "prec"
Engine and server modified to actually trigger calcs.
Working on reactivity.

* FWI87 and FWI25 now react to all input changes?

* Fixes #9

* Tool tips and help galore!

* Some minor changes.

* Fix issue 22 29 (#31)

* 1. Adjusts the tables in UI to be prettier.
2. Fixes for files in reverse chronological order:
- filters work now
- hFWI works now
- works on both FWI87 and FWI25
side effect of the fix is to that data is reordered. If needed we can mark for sorting back into reverse chrono order for user.

* Now skips blank rows and rows that have only delimiters.

* Some style and control updates to make it prettier.

* Fixes #22 (#34)

-Horizontal slider was always there but is now easier to notice and see since the card is sized properly.
-Decimal places are now configurable when deployed. We have some defaults we can tweak.
-Column headers line up.

* Fix issue 14 18 (#38)

* - Fixes 14 and 18 by adding in inputs for solar radiation, date and time fields.
- Fixes 32 by adding Multistation support for plots and tables supported for fwi87 and fwi25
- TODO: fix timezones for sunrise,sunset and other timestamp fields

* Finally sorted out all the timezone bs.
ToDO:
Check plot data selection.
Sort out daily fwi 12:00 stuff (issue 19)

* AH! Fixed the times for real now. What a nightmare. Why can't things just be timezone aware or all in UTC.

* Okay, now for real, times are sorted.
The plots work as I expect, the tables look nice and so does the modal, and the helpers and tool tips.
But now it's super slow so will try to look into that.
Fixes #9
Fixes #14
Fixes #18
Fixes #12
Fixes #24
I think it fixes number 11,  27 but not sure.

* Hmm, can't decide if this is better or not but pushing it up.
Might be better to revert the plot code. But if not we need to fix the faceting and labelling.
Otherwise the next thing to work on is how slow the app is and that there's no indication that the button was pressed and it's now thinking.

* Everything but the timing of the spinners now.

* Timing fixed!

* Getting closer

* Finally found where the warning was coming from in the tables.

* Update to latest version of ngcffdrs main branch
Repair to ng/util.r is_sequential function to properly test for sequential data.
Closes #36.